Support dynamic filename filter for helm-do-ag
The idea is to parse a pattern like "asdf !el jkl !txt" so that the ag
command becomes -Gel|txt asdf jkl. The ! terms are added to the -G
argument dynamically, so that if you are in the middle of a search with
a lot of results you can quickly narrow by filename without starting
over by just adding ! terms.

The feature is off by default and controlled by the variable
`helm-do-ag--allow-filename-filter`.