Remove "--" from construct-do-ag-command
This extra `"--"` breaks the functionality of the `-g` flag which allows ag to search file names (rather than the contents of the file).

Before this change, commands (like those which come from `helm-projectile-ag`) create lists such that observation of `helm-ag--last-command` resembles the following:

    ... "--ignore" "_darcs" "--ignore" "{arch}" "-g" "--" "test")

With this, ag returns an error. However, without the additional list item `"--"`, the command executes as follows, returning all files with names matching the pattern "test".