Make helm-ag--command-feature(s) into a list
This commit replaces helm-ag--command-feature with
helm-ag--command-features, a list that can properly express the features
available from the chosen command and its parameters.

The list always includes the name of the command for direct reference
(for example 'pt), but also includes flags such as 'fixed, 're2 or
'pcre.

Along with supporting the existing command set better, this also adds
support for ripgrep by recognizing its limitation to re2-compatible
regular expressions (similar to pt).