Change the directory of a reused temp buffer
When `helm-ag-use-temp-buffer` is set to t, previewing a search result's
file (<kbd>C-j</kbd>) won't open the file and leave it open afterwards,
but will show the contents of the file in a temporary buffer.

However, consider the following scenario: you first preview search
results of a search initiated in some directory `/a`, and then try to
preview search results of a search initiated in some directory `/b`.
When previewing the search results in some file `x` in `/b`, you will
get the following error: "Opening input file: No such file or directory,
/a/x".

This is because in `helm-ag--open-file-with-temp-buffer`, the temporary
buffer is created only once. The buffer-local variable
`default-directory` of this buffer is implicitly set to the directory
that initiated the first search (`/a` in our example). This value is
implicitly used in the function call `(insert-file-contents filename)`,
where `filename` is a relative filename that will be expanded in
`default-directory`.

The next time `helm-ag--open-file-with-temp-buffer` is called, the
temporary buffer along with its value for `default-directory` (`/a`) is
reused. When `(insert-file-contents "x")` is then called, we get the
error that `/a/x` is missing.

The fix is simple: change the temporary buffer's value of
`default-directory` to the directory in which the current search is
initiated.