Always update status if search process ends
If the search process ends with a non-zero exit code, the helm status
was never updated. E.g. Grep, Ripgrep and likely more use exit code 1 to
indicate that nothing is found, a frequently occurring scenario that
left the user wondering if the process is still searching or not.

Now it will always update status if the process is not live anymore.